REST API Kullanımı/Yapay Zeka Entegrasyonu Yapıldı

+++ Controllers/AntrenorlerController.cs
@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SporSalonuYonetimSistemi.Data;
 using SporSalonuYonetimSistemi.Models;
-using Microsoft.AspNetCore.Authorization; // Bunu ekle
 
 namespace SporSalonuYonetimSistemi.Controllers
 {
+    // DÜZELTME 1: Sadece [Authorize] yaptık. Böylece Üyeler de girebilir.
     [Authorize]
     public class AntrenorlerController : Controller
     {
@@ -21,83 +17,74 @@ namespace SporSalonuYonetimSistemi.Controllers
             _context = context;
         }
 
-        // GET: Antrenorler
+        // Herkes Listeyi Görebilir
         public async Task<IActionResult> Index()
         {
             return View(await _context.Antrenorler.ToListAsync());
         }
 
-        // GET: Antrenorler/Details/5
+        // Herkes Detayları Görebilir
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var antrenor = await _context.Antrenorler
-                .FirstOrDefaultAsync(m => m.AntrenorId == id);
-            if (antrenor == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var antrenor = await _context.Antrenorler.FirstOrDefaultAsync(m => m.AntrenorId == id);
+            if (antrenor == null) return NotFound();
             return View(antrenor);
         }
 
-        // GET: Antrenorler/Create
+        // DÜZELTME 2: Ekleme işlemini SADECE ADMIN yapabilir
         [Authorize(Roles = "Admin")]
         public IActionResult Create()
         {
+            ViewBag.Hizmetler = _context.Hizmetler.ToList();
             return View();
         }
 
-        // POST: Antrenorler/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> Create([Bind("AntrenorId,Ad,Soyad,UzmanlikAlanlari,FotografUrl")] Antrenor antrenor)
+        public async Task<IActionResult> Create([Bind("AntrenorId,Ad,Soyad,UzmanlikAlanlari,FotografUrl")] Antrenor antrenor, int[] selectedHizmetler)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(antrenor);
                 await _context.SaveChangesAsync();
+
+                if (selectedHizmetler != null)
+                {
+                    foreach (var hizmetId in selectedHizmetler)
+                    {
+                        _context.AntrenorHizmetleri.Add(new AntrenorHizmet { AntrenorId = antrenor.AntrenorId, HizmetId = hizmetId });
+                    }
+                    await _context.SaveChangesAsync();
+                }
                 return RedirectToAction(nameof(Index));
             }
+            ViewBag.Hizmetler = _context.Hizmetler.ToList();
             return View(antrenor);
         }
 
-        // GET: Antrenorler/Edit/5
+        // DÜZELTME 3: Düzenleme işlemini SADECE ADMIN yapabilir
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var antrenor = await _context.Antrenorler.FindAsync(id);
-            if (antrenor == null)
-            {
-                return NotFound();
-            }
+            if (antrenor == null) return NotFound();
+
+            ViewBag.Hizmetler = _context.Hizmetler.ToList();
+            ViewBag.SeciliHizmetler = await _context.AntrenorHizmetleri
+                .Where(ah => ah.AntrenorId == id).Select(ah => ah.HizmetId).ToListAsync();
+
             return View(antrenor);
         }
 
-        // POST: Antrenorler/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin")]
-        public async Task<IActionResult> Edit(int id, [Bind("AntrenorId,Ad,Soyad,UzmanlikAlanlari,FotografUrl")] Antrenor antrenor)
+        public async Task<IActionResult> Edit(int id, [Bind("AntrenorId,Ad,Soyad,UzmanlikAlanlari,FotografUrl")] Antrenor antrenor, int[] selectedHizmetler)
         {
-            if (id != antrenor.AntrenorId)
-            {
-                return NotFound();
-            }
+            if (id != antrenor.AntrenorId) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -105,54 +92,50 @@ namespace SporSalonuYonetimSistemi.Controllers
                 {
                     _context.Update(antrenor);
                     await _context.SaveChangesAsync();
+
+                    var eskiler = _context.AntrenorHizmetleri.Where(ah => ah.AntrenorId == id);
+                    _context.AntrenorHizmetleri.RemoveRange(eskiler);
+                    await _context.SaveChangesAsync();
+
+                    if (selectedHizmetler != null)
+                    {
+                        foreach (var hizmetId in selectedHizmetler)
+                        {
+                            _context.AntrenorHizmetleri.Add(new AntrenorHizmet { AntrenorId = id, HizmetId = hizmetId });
+                        }
+                        await _context.SaveChangesAsync();
+                    }
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!AntrenorExists(antrenor.AntrenorId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!AntrenorExists(antrenor.AntrenorId)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
+            ViewBag.Hizmetler = _context.Hizmetler.ToList();
+            ViewBag.SeciliHizmetler = await _context.AntrenorHizmetleri
+                .Where(ah => ah.AntrenorId == id).Select(ah => ah.HizmetId).ToListAsync();
             return View(antrenor);
         }
 
-        // GET: Antrenorler/Delete/5
+        // DÜZELTME 4: Silme işlemini SADECE ADMIN yapabilir
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var antrenor = await _context.Antrenorler
-                .FirstOrDefaultAsync(m => m.AntrenorId == id);
-            if (antrenor == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var antrenor = await _context.Antrenorler.FirstOrDefaultAsync(m => m.AntrenorId == id);
+            if (antrenor == null) return NotFound();
             return View(antrenor);
         }
 
-        // POST: Antrenorler/Delete/5
+        [Authorize(Roles = "Admin")]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var antrenor = await _context.Antrenorler.FindAsync(id);
-            if (antrenor != null)
-            {
-                _context.Antrenorler.Remove(antrenor);
-            }
-
+            if (antrenor != null) _context.Antrenorler.Remove(antrenor);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
